requirements.txt
# Core Django
asgiref==3.10.0
Django==5.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
pytz==2025.2

# Database
mssql-django==1.6
pyodbc==5.3.0

# Configuration
python-decouple==3.8

# Forms
django-crispy-forms==2.5
crispy-bootstrap5==2025.6

# Images (para validación de URLs, aunque no se suben archivos)
pillow==12.0.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0  # Para servir archivos estáticos en producción
